+++ selene/support/shared/__init__.py
@@ -36,24 +36,11 @@ from selene.support.shared.config import SharedConfig
 
 config = SharedConfig()
 
-
-def _take_screenshot():
-    path = config.reports_folder
-    next_id = next(config.counter)
-    filename = f'screen_{next_id}'
-    driver = config.driver
-    screenshot_path = os.path.join(path, f'{filename}.png')
-    folder = os.path.dirname(screenshot_path)
-
-    if not os.path.exists(folder):
-        os.makedirs(folder)
-
-    # todo: refactor to catch errors smartly in get_screenshot_as_file
-    return screenshot_path if driver.get_screenshot_as_file(screenshot_path) else 'None'
+browser = SharedBrowser(config)
 
 
 def _save_and_log_screenshot(error: TimeoutException) -> Exception:
-    path = _take_screenshot()
+    path = browser.save_screenshot()
     return TimeoutException(
         error.msg + f'''
 Screenshot: file://{path}''',
@@ -61,55 +48,22 @@ Screenshot: file://{path}''',
         error.stacktrace)
 
 
-def _take_html_dump():
-    path = config.reports_folder
-    next_id = next(config.counter)
-    filename = f'source_{next_id}'
-    driver: WebDriver = config.driver
-    page_source_path = os.path.join(path, f'{filename}.html')
-    folder = os.path.dirname(page_source_path)
-
-    if not os.path.exists(folder):
-        os.makedirs(folder)
-
-    # todo: refactor to catch errors smartly in get_screenshot_as_file
-
-    # todo: consider moving this to browser query._page_source_as_file(filename)
-    def _get_page_source_as_file(filename) -> bool:
-        if not filename.lower().endswith('.html'):
-            warnings.warn("name used for saved pagesource does not match file "
-                          "type. It should end with an `.html` extension", UserWarning)
-
-        html = driver.page_source
-        try:
-            with open(filename, 'w') as f:
-                f.write(html)
-        except IOError:
-            return False
-        finally:
-            del html
-        return True
-
-    return page_source_path if _get_page_source_as_file(page_source_path) else 'None'
-
-
-def _save_and_log_html_dump(error: TimeoutException) -> Exception:
-    path = _take_html_dump()
+def _save_and_log_page_source(error: TimeoutException) -> Exception:
+    filename = browser.latest_screenshot.replace('.png', '.html') if browser.latest_screenshot else None
+    path = browser.save_page_source(filename)
     return TimeoutException(
         error.msg + f'''
-Page Source: file://{path}''',
+PageSource: file://{path}''',
         error.screen,
         error.stacktrace)
 
 
-def _compose(*functions):
-    return functools.reduce(lambda f, g: lambda x: f(g(x)), functions, lambda x: x)
+def _pipe(*functions):
+    return functools.reduce(lambda f, g: lambda x: f(g(x)), functions[::-1], lambda x: x)
 
 
 # todo: consider more flat style: Hooks(wait_failure=_save_and_log_screenshot)
 # todo: consider making screenshots configurable (turn on/off)
-config.hooks = Hooks(wait=WaitHooks(failure=_compose(
-    _save_and_log_html_dump,
-    _save_and_log_screenshot)))
-
-browser = SharedBrowser(config)
+config.hooks = Hooks(wait=WaitHooks(failure=_pipe(
+    _save_and_log_screenshot,
+    _save_and_log_page_source)))