+++ selene/support/shared/browser.py
@@ -29,12 +29,12 @@ from selene.core.entity import Browser, Collection
 from selene.core.configuration import Config
 from selene.common.none_object import NoneObject
 from selene.support.shared.config import SharedConfig
-from selene.support.shared import config
 
 
 class SharedBrowser(Browser):
     def __init__(self, config: SharedConfig):
         self._latest_screenshot = NoneObject('selene.SharedBrowser._latest_screenshot')
+        self._latest_page_source = NoneObject('selene.SharedBrowser._latest_page_source')
         super().__init__(config)
 
     @property
@@ -58,6 +58,75 @@ class SharedBrowser(Browser):
 
         return webdriver
 
+    def _next_generated_absolute_filename(self, prefix='', suffix=''):
+        path = self.config.reports_folder
+        next_id = next(self.config.counter)
+        filename = f'{prefix}{next_id}{suffix}'
+        file = os.path.join(path, f'{filename}')
+
+        folder = os.path.dirname(file)
+        if not os.path.exists(folder):
+            os.makedirs(folder)
+
+        return file
+
+    def save_screenshot(self, file: str = None):
+        warnings.warn('might be deprecated', FutureWarning)
+
+        if not file:
+            file = self._next_generated_absolute_filename(suffix='.png')
+
+        # todo: refactor to catch errors smartly in get_screenshot_as_file
+        self._latest_screenshot = file if self.driver.get_screenshot_as_file(file) else None
+        return self._latest_screenshot
+
+    @property
+    def latest_screenshot(self) -> str:
+        warnings.warn('might be deprecated', FutureWarning)
+
+        class CallableString(str):
+            def __new__(cls, value):
+                obj = str.__new__(cls, value)
+                return obj
+
+            def __call__(self, *args, **kwargs):
+                warnings.warn('browser.latest_screenshot() is deprecated, '
+                              'use browser.latest_screenshot as a property. ',
+                              DeprecationWarning)
+                return self[:]
+
+        return CallableString(self._latest_screenshot)
+
+    # todo: consider moving this to browser command.save_page_source(filename)
+    def save_page_source(self, file: str = None):
+        warnings.warn('might be deprecated', FutureWarning)
+
+        if not file:
+            file = self._next_generated_absolute_filename(suffix='.html')
+
+        if not file.lower().endswith('.html'):
+            warnings.warn("name used for saved pagesource does not match file "
+                          "type. It should end with an `.html` extension", UserWarning)
+
+        html = self.driver.page_source
+
+        try:
+            with open(file, 'w') as f:
+                f.write(html)
+        except IOError:
+            self._latest_page_source = None
+        finally:
+            del html
+
+        self._latest_page_source = file
+
+        return file
+
+    @property
+    def latest_page_source(self):
+        warnings.warn('might be deprecated', FutureWarning)
+        return self._latest_page_source
+
     def quit_driver(self):
         warnings.warn('deprecated; use browser.quit() instead', DeprecationWarning)
         self.quit()
@@ -80,30 +149,6 @@ class SharedBrowser(Browser):
         warnings.warn('use browser.all instead', DeprecationWarning)
         return self.all(css_or_xpath_or_by)
 
-    def take_screenshot(self, path=None, filename=None):
-        warnings.warn('might be deprecated', PendingDeprecationWarning)
-        if not path:
-            path = self.config.reports_folder
-        if not filename:
-            id = next(config.counter)
-            filename = f'screen_{id}'
-
-        screenshot_path = os.path.join(path,f'{filename}.png')
-
-        folder = os.path.dirname(screenshot_path)
-        if not os.path.exists(folder):
-            os.makedirs(folder)
-
-        self.driver.save_screenshot(screenshot_path)
-
-        self._latest_screenshot = screenshot_path
-
-        return screenshot_path
-
-    def latest_screenshot(self):
-        warnings.warn('might be deprecated', PendingDeprecationWarning)
-        return self._latest_screenshot
-
     def wait_to(self, webdriver_condition, timeout=None, polling=None):
         warnings.warn('use browser.should instead', DeprecationWarning)
         tuned_self = self if timeout is None else self.with_(Config(timeout=timeout))
@@ -117,3 +162,8 @@ class SharedBrowser(Browser):
     def title(self):
         warnings.warn('use browser.driver.title or browser.get(query.title) instead', DeprecationWarning)
         return self.driver.title
+
+    def take_screenshot(self, path=None, filename=None):
+        warnings.warn('deprecated; use browser.save_screenshot(filename) instead', DeprecationWarning)
+        # todo: refactor to deal with cases when path is not set but filename is set
+        return self.save_screenshot(((path or '') + (filename or '')) or None)